requirements.txt
python-tgpt
nest-asyncio
g4f
discord.py
